% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_RCondaRun.R
\name{RCondaRun}
\alias{RCondaRun}
\title{Run Python function in a conda environment}
\usage{
RCondaRun(func, args, env_name)
}
\arguments{
\item{func}{function}

\item{args}{list of argument values to \code{func}}

\item{env_name}{name of conda environment}
}
\description{
Uses Python, \code{reticulate} and \code{callr} to run a function \code{func} with arguments \code{args} with a specific conda environment activated.
This function can use \code{source_python}, \code{py_eval} and other \code{reticulate} functions to run Python code.
The function is run in a separate R session, because \code{reticulate} cannot switch between conda environments in a single R session.
}
\details{
This function does not attach the \code{reticulate} namespace in your R session.
}
